requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
amqp==2.2.2               # via kombu
billiard==3.5.0.3         # via celery
boto3==1.7.19
botocore==1.10.19         # via boto3, s3transfer
celery[redis]==4.1.0
click==6.7                # via flask
docutils==0.14            # via botocore
flask==1.0.2
itsdangerous==0.24        # via flask
jinja2==2.10              # via flask
jmespath==0.9.3           # via boto3, botocore
kombu==4.1.0              # via celery
markupsafe==1.0           # via jinja2
pyee==5.0.0               # via pyppeteer
pyppeteer==0.0.17
python-dateutil==2.7.2    # via botocore
python-dotenv==0.8.2
pytz==2018.4              # via celery
redis==2.10.6             # via celery
s3transfer==0.1.13        # via boto3
six==1.11.0               # via python-dateutil
vine==1.1.4               # via amqp
websockets==4.0.1         # via pyppeteer
werkzeug==0.14.1          # via flask
